Relationships

+++ src/idea/idea.service.ts
@@ -2,9 +2,9 @@ import {HttpException, HttpStatus, Injectable} from '@nestjs/common';
 import {InjectRepository} from '@nestjs/typeorm';
 import {getRepository, Repository} from 'typeorm';
 
-import {IdeaDTO} from './idea.dto';
+import {IdeaDTO, IdeaRO} from './idea.dto';
 import {IdeaEntity} from './idea.entity';
-import {UserEntity} from "../user/user.entity";
+import {UserEntity} from '../user/user.entity';
 
 @Injectable()
 export class IdeaService {
@@ -15,35 +15,45 @@ export class IdeaService {
     private userRepository: Repository<UserEntity>
   ) {}
 
-  private toResponseObject(idea: IdeaEntity) {
+  private toResponseObject(idea: IdeaEntity): IdeaRO {
     return {
       ...idea,
       user: idea.user.toResponseObject(false)
     }
   }
 
-  async showAllIdeas() {
+  private ensureOwnership(
+    idea: IdeaEntity,
+    userId: string
+  ) {
+    if (idea.user.id !== userId) {
+      throw new HttpException(
+        'Incorrect user',
+        HttpStatus.UNAUTHORIZED
+      );
+    }
+  }
+
+  async showAllIdeas(): Promise<IdeaRO[]> {
     try {
-      const queryBuilder = await getRepository(IdeaEntity)
-        .createQueryBuilder('idea')
-        .leftJoinAndSelect('idea.user', 'user');
-      queryBuilder.where('1 = 1');
-      queryBuilder.orderBy('idea.idea', 'DESC');
-      const ideasCount = await queryBuilder.getCount();
-      const ideas = await queryBuilder.getMany();
-      return { ideasCount, ideas };
-      // const ideas = await this.ideaRepository.find({
-      //   relations: ['user']
-      // });
-      // return ideas.map(idea => {
-      //   console.log(idea);
-      // });
+      // const queryBuilder = await getRepository(IdeaEntity)
+      //   .createQueryBuilder('idea')
+      //   .leftJoinAndSelect('idea.user', 'user');
+      // queryBuilder.where('1 = 1');
+      // queryBuilder.orderBy('idea.idea', 'DESC');
+      // const ideasCount = await queryBuilder.getCount();
+      // const ideas = await queryBuilder.getMany();
+      // return { ideasCount, ideas };
+      const ideas = await this.ideaRepository.find({
+        relations: ['user']
+      });
+      return ideas.map(idea => this.toResponseObject(idea));
     } catch (err) {
       throw new HttpException(`Bad query ${err}`, HttpStatus.BAD_REQUEST)
     }
   }
 
-  async createIdea(userId: string, data: IdeaDTO) {
+  async createIdea(userId: string, data: IdeaDTO): Promise<IdeaRO> {
     const user = await this.userRepository.findOne({
       where: {
         id: userId
@@ -53,44 +63,52 @@ export class IdeaService {
       ...data,
       user: user
     });
-    console.log({
-      ...data,
-      user: user
-    });
     await this.ideaRepository.save(idea);
     return this.toResponseObject(idea);
   }
 
-  async readIdea(id: string) {
+  async readIdea(id: string): Promise<IdeaRO> {
     const idea = await this.ideaRepository.findOne({
-      where: { id }
+      where: { id },
+      relations: ['user']
     });
     if (!idea) {
       throw new HttpException('Not found', HttpStatus.NOT_FOUND);
     }
-    return idea;
+    return this.toResponseObject(idea);
   }
 
-  async updateIdea(id: string, data: Partial<IdeaDTO>) {
+  async updateIdea(
+    id: string,
+    userId: string,
+    data: Partial<IdeaDTO>
+  ): Promise<IdeaRO> {
     let idea = await this.ideaRepository.findOne({
-      where: { id }
+      where: { id },
+      relations: ['user']
     });
     if (!idea) {
       throw new HttpException('Not found', HttpStatus.NOT_FOUND);
     }
+    this.ensureOwnership(idea, userId);
     await this.ideaRepository.update({ id }, data);
-    idea = await this.ideaRepository.findOne({ where: { id } });
-    return idea;
+    idea = await this.ideaRepository.findOne({
+      where: { id },
+      relations: ['user']
+    });
+    return this.toResponseObject(idea);
   }
 
-  async destroyIdea(id: string) {
+  async destroyIdea(id: string, userId: string,) {
     const idea = await this.ideaRepository.findOne({
-      where: { id }
+      where: { id },
+      relations: ['user']
     });
     if (!idea) {
       throw new HttpException('Not found', HttpStatus.NOT_FOUND);
     }
+    this.ensureOwnership(idea, userId);
     await this.ideaRepository.delete({ id });
-    return idea;
+    return this.toResponseObject(idea);
   }
 }
